backend/requirements.txt
Django==5.2.8
djangorestframework==3.15.2
django-cors-headers==4.6.0
channels>=4.0
daphne>=4.0
dashscope>=1.25.3
aiohttp>=3.9.0
websockets>=12.0
groq>=0.11.0
Pillow==11.0.0
python-dotenv==1.2.1
openai==1.58.1
requests==2.32.3
gunicorn==21.2.0
gevent==24.2.1
